refactor: make separate conditional approval object field template
