+++ app/components/Analyst/Tracking/ConditionalApproval/ConditionalApprovalTheme.ts
@@ -1,0 +1,10 @@
+import { ThemeProps } from '@rjsf/core';
+import TrackingTheme from '../TrackingTheme';
+import ConditionalApprovalObjectFieldTemplate from './ConditionalApprovalObjectFieldTemplate';
+
+const ConditionalApprovalReadOnlyTheme: ThemeProps = {
+  ...TrackingTheme,
+  ObjectFieldTemplate: ConditionalApprovalObjectFieldTemplate,
+};
+
+export default ConditionalApprovalReadOnlyTheme;
