+++ app/components/Analyst/Tracking/ConditionalApproval/index.ts
@@ -1,5 +1,6 @@
 export { default as BCHeader } from './BCHeader';
 export { default as ConditionalApprovalReadOnlyTheme } from './ConditionalApprovalReadOnlyTheme';
 export { default as ConditionalApprovalModal } from './ConditionalApprovalModal';
+export { default as ConditionalApprovalTheme } from './ConditionalApprovalTheme';
 export { default as ISEDHeader } from './ISEDHeader';
 export { default as StyledStatus } from './StyledStatus';
